service/requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
boto3==1.34.121
numpy==1.26.4
scikit-learn==1.2.2
python-dotenv==1.0.0
pandas==2.2.2